requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.5
    # via
    #   discord-py
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
dataclasses-json==0.5.13
    # via langchain
discord-py==2.3.1
    # via -r requirements.in
faiss-cpu==1.7.4
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   requests
    #   yarl
langchain==0.0.242
    # via
    #   -r requirements.in
    #   promptlayer
langsmith==0.0.15
    # via langchain
marshmallow==3.20.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numexpr==2.8.4
    # via langchain
numpy==1.25.1
    # via
    #   -r requirements.in
    #   langchain
    #   numexpr
openai==0.27.8
    # via -r requirements.in
openapi-schema-pydantic==1.2.4
    # via langchain
packaging==23.1
    # via marshmallow
promptlayer==0.1.93
    # via -r requirements.in
pydantic==1.10.12
    # via
    #   langchain
    #   langsmith
    #   openapi-schema-pydantic
python-dotenv==1.0.0
    # via -r requirements.in
python-magic==0.4.27
    # via -r requirements.in
pyyaml==6.0.1
    # via langchain
regex==2023.6.3
    # via tiktoken
requests==2.31.0
    # via
    #   langchain
    #   langsmith
    #   openai
    #   promptlayer
    #   tiktoken
sqlalchemy==2.0.19
    # via langchain
tenacity==8.2.2
    # via langchain
tiktoken==0.4.0
    # via -r requirements.in
tqdm==4.65.0
    # via openai
typing-extensions==4.7.1
    # via
    #   pydantic
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.0.4
    # via requests
yarl==1.9.2
    # via aiohttp
